+++ app/src/main/kotlin/com/simplemobiletools/clock/activities/SplashActivity.kt
@@ -16,7 +16,7 @@ class SplashActivity : BaseSplashActivity() {
             intent.extras?.containsKey(OPEN_TAB) == true -> {
                 Intent(this, MainActivity::class.java).apply {
                     putExtra(OPEN_TAB, intent.getIntExtra(OPEN_TAB, TAB_CLOCK))
-                    putExtra(TIMER_ID, intent.getLongExtra(TIMER_ID, INVALID_TIMER_ID))
+                    putExtra(TIMER_ID, intent.getIntExtra(TIMER_ID, INVALID_TIMER_ID))
                     startActivity(this)
                 }
             }
