+++ app/src/main/kotlin/com/simplemobiletools/clock/models/Timer.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "timers")
 data class Timer(
-    @PrimaryKey(autoGenerate = true) var id: Long?,
+    @PrimaryKey(autoGenerate = true) var id: Int?,
     var seconds: Int,
     val state: TimerState,
     var vibrate: Boolean,
